Amd Slingshot Hackathon project/requirements.txt
flask
pandas
scikit-learn
joblib
openai
python-dotenv
gunicorn
